refactor : split test_api in model classes
